+++ src/Game/Text/el/HelpText.el.sjson.lua
@@ -36,16 +36,32 @@ local newData = {
 		DisplayName = "ΤΑΡΤΑΡΑ (Εφιάλτης)",
 	},
 	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Tartarus_Short",
+		DisplayName = "Ταρταρα",
+	},
+	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Asphodel",
 		DisplayName = "ΑΣΦΟΔΕΛΟΣ (Εφιάλτης)",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Asphodel_Short",
+		DisplayName = "Ασφοδελος",
 	},
 	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Elysium",
 		DisplayName = "ΕΛΥΣΙΟΥΜ (Εφιάλτης)",
 	},
 	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Elysium_Short",
+		DisplayName = "Ελυσιουμ",
+	},
+	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Styx",
 		DisplayName = "ΣΤΥΞ (Εφιάλτης)",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Styx_Short",
+		DisplayName = "Στυξ",
 	},
 	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Surface",
@@ -53,8 +69,16 @@ local newData = {
 		DisplayName = "ΕΠΙΦΑΝΕΙΑ (Εφιάλτης)",
 	},
 	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Surface_Short",
+		DisplayName = "Επιφανεια",
+	},
+	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Challenge",
 		DisplayName = "ΕΡΕΒΟΣ (Εφιάλτης)",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Challenge_Short",
+		DisplayName = "Ερεβος",
 	},
 	-- #endregion
 
@@ -1043,7 +1067,41 @@ local newData = {
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_UnlockPostBossGiftRackIncantation_Flavor",
 	-- 	Description =
-	-- 	"As Princess of Nightmares, no dream is too far removed from reality to not allow yourself to be reminded of those you hold dear."
+	-- 	"For the Princess of Nightmares, no dream is too far removed from reality to not allow her to be reminded of those she holds dear."
+	-- },
+	-- Add Reprieve/Fountain Chambers
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockTartarusReprieveIncantation",
+	-- 	DisplayName = "Eerie Lifespring",
+	-- 	Description =
+	-- 	"Reveal a hidden chamber in {$Keywords.ModsNikkelMHadesBiomesLocation_Hades_Tartarus_Short} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}, where you may restore some {!Icons.HealthRestoreHome} and elude your foes."
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockTartarusReprieveIncantation_Flavor",
+	-- 	Description =
+	-- 	"Tucked away in a long lost corner of Tartarus is an abandoned bath house, where the Erinyes once sought respite from their eternal duties.",
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockAsphodelReprieveIncantation",
+	-- 	DisplayName = "Scalding Lifespring",
+	-- 	Description =
+	-- 	"Reveal a hidden chamber in {$Keywords.ModsNikkelMHadesBiomesLocation_Hades_Asphodel_Short} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}, where you may restore some {!Icons.HealthRestoreHome} and elude your foes."
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockAsphodelReprieveIncantation_Flavor",
+	-- 	Description =
+	-- 	"How this forgotten fountain has not yet evaporated in the heat of Asphodel is a mystery best not pondered too deeply, lest it leads to answers you do not wish to know.",
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockElysiumReprieveIncantation",
+	-- 	DisplayName = "Foggy Lifespring",
+	-- 	Description =
+	-- 	"Reveal a hidden chamber in {$Keywords.ModsNikkelMHadesBiomesLocation_Hades_Elysium_Short} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}, where you may restore some {!Icons.HealthRestoreHome} and elude your foes."
+	-- },
+	-- {
+	-- 	Id = "ModsNikkelMHadesBiomes_UnlockElysiumReprieveIncantation_Flavor",
+	-- 	Description =
+	-- 	"Even the greatest of heroes who have passed into Elysium are not allowed to drink of the waters flowing from it's fountains, which is reserved for the gods alone.",
 	-- },
 	-- #endregion