+++ src/Game/Text/en/HelpText.en.sjson.lua
@@ -54,24 +54,48 @@ local newData = {
 		DisplayName = "TARTARUS (Nightmare)",
 	},
 	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Tartarus_Short",
+		DisplayName = "Tartarus",
+	},
+	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Asphodel",
 		DisplayName = "ASPHODEL (Nightmare)",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Asphodel_Short",
+		DisplayName = "Asphodel",
 	},
 	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Elysium",
 		DisplayName = "ELYSIUM (Nightmare)",
 	},
 	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Elysium_Short",
+		DisplayName = "Elysium",
+	},
+	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Styx",
 		DisplayName = "STYX (Nightmare)",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Styx_Short",
+		DisplayName = "Styx",
 	},
 	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Surface",
 		DisplayName = "GREECE (Nightmare)",
 	},
 	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Surface_Short",
+		DisplayName = "Greece",
+	},
+	{
 		Id = "ModsNikkelMHadesBiomesLocation_Hades_Challenge",
 		DisplayName = "EREBUS (Nightmare)",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomesLocation_Hades_Challenge_Short",
+		DisplayName = "Erebus",
 	},
 	-- #endregion
 
@@ -1065,7 +1089,41 @@ local newData = {
 	{
 		Id = "ModsNikkelMHadesBiomes_UnlockPostBossGiftRackIncantation_Flavor",
 		Description =
-		"As Princess of Nightmares, no dream is too far removed from reality to not allow yourself to be reminded of those you hold dear."
+		"For the Princess of Nightmares, no dream is too far removed from reality to not allow her to be reminded of those she holds dear."
+	},
+	-- Add Reprieve/Fountain Chambers
+	{
+		Id = "ModsNikkelMHadesBiomes_UnlockTartarusReprieveIncantation",
+		DisplayName = "Eerie Lifespring",
+		Description =
+		"Reveal a hidden chamber in {$Keywords.ModsNikkelMHadesBiomesLocation_Hades_Tartarus_Short} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}, where you may restore some {!Icons.HealthRestoreHome} and elude your foes."
+	},
+	{
+		Id = "ModsNikkelMHadesBiomes_UnlockTartarusReprieveIncantation_Flavor",
+		Description =
+		"Tucked away in a long lost corner of Tartarus is an abandoned bath house, where the Erinyes once sought respite from their eternal duties.",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomes_UnlockAsphodelReprieveIncantation",
+		DisplayName = "Scalding Lifespring",
+		Description =
+		"Reveal a hidden chamber in {$Keywords.ModsNikkelMHadesBiomesLocation_Hades_Asphodel_Short} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}, where you may restore some {!Icons.HealthRestoreHome} and elude your foes."
+	},
+	{
+		Id = "ModsNikkelMHadesBiomes_UnlockAsphodelReprieveIncantation_Flavor",
+		Description =
+		"How this forgotten fountain has not yet evaporated in the heat of Asphodel is a mystery best not pondered too deeply, lest it leads to answers you do not wish to know.",
+	},
+	{
+		Id = "ModsNikkelMHadesBiomes_UnlockElysiumReprieveIncantation",
+		DisplayName = "Foggy Lifespring",
+		Description =
+		"Reveal a hidden chamber in {$Keywords.ModsNikkelMHadesBiomesLocation_Hades_Elysium_Short} in the {$Keywords.ModsNikkelMHadesBiomesModdedRoute}, where you may restore some {!Icons.HealthRestoreHome} and elude your foes."
+	},
+	{
+		Id = "ModsNikkelMHadesBiomes_UnlockElysiumReprieveIncantation_Flavor",
+		Description =
+		"Even the greatest of heroes who have passed into Elysium are not allowed to drink of the waters flowing from it's fountains, which is reserved for the gods alone.",
 	},
 	-- #endregion