+++ src/Scripts/KeywordData.lua
@@ -1,5 +1,11 @@
 local newKeywords = {
 	"ModsNikkelMHadesBiomesModdedRoute",
+	"ModsNikkelMHadesBiomesLocation_Hades_Tartarus_Short",
+	"ModsNikkelMHadesBiomesLocation_Hades_Asphodel_Short",
+	"ModsNikkelMHadesBiomesLocation_Hades_Elysium_Short",
+	"ModsNikkelMHadesBiomesLocation_Hades_Styx_Short",
+	"ModsNikkelMHadesBiomesLocation_Hades_Surface_Short",
+	"ModsNikkelMHadesBiomesLocation_Hades_Challenge_Short",
 	"ShrinePointGate",
 }
 game.ConcatTableValuesIPairs(game.KeywordList, newKeywords)